+++ index.js
@@ -50,7 +50,7 @@ const safeObject = require('./util/safe_object');
  * @property {false} [fence_pre_attrs] Disable fence pre attributes.
  * @property {false|import('./modifiers/fence_classes').FenceClassesOptions} [fence_classes] Disable fence class filtering, or set options for the feature.
  * @property {false|import('./modifiers/heading_id').HeadingIdOptions} [heading_id] Disable Ids on headings, or set options for the feature.
- * @property {false|import('./modifiers/image_size').ImageSizeOptions} [image_size] Disable image size syntax, or set options for the feature.
+ * @property {false|import('./modifiers/image_settings').ImageSettingsOptions} [image_settings] Disable image settings syntax, or set options for the feature.
  * @property {false|import('./modifiers/prismjs').PrismJsOptions} [prismjs] Disable Prism highlighting, or set options for the feature.
  */
 
@@ -173,8 +173,8 @@ module.exports = (md, options) => {
         md.use(require('./modifiers/heading_id'), safeObject(optsObj.heading_id));
     }
 
-    if (optsObj.image_size !== false) {
-        md.use(require('./modifiers/image_size'), safeObject(optsObj.image_size));
+    if (optsObj.image_settings !== false) {
+        md.use(require('./modifiers/image_settings'), safeObject(optsObj.image_settings));
     }
 
     if (optsObj.prismjs !== false) {